Fix numerical precision issues when running on a 8-bit PIC
This fixes a problem when running on an 8-bit PIC where the calcuations used in interplotation are not carried out using a 32-bit signed integer as required but as a 16-bit unsigned value. Explicitly casting the first use in each expression ensures that the correct type is used for the intermediate value.

Tests:
 - Run updated unit test with a real-world test value
 - Build and run on a PIC and verify in the debugger that the
   calculation is operating the same as the unit test for the
   same input value

+++ lib/mapper.c
@@ -37,15 +37,19 @@ static uint16_t InterpolateBinValue(
     const uint16_t* outputMap )
 {
     //
-    // Use int32_t values to differences to avoid overflowing a uin16_t when
+    // Use int32_t values for differences to avoid overflowing a uin16_t when
     // multiplying on an 8-bit PIC
+    //
+    // RHS variables need to be cast to int32_t to ensure calculations are
+    // carried out in 32-bits
     //
     // Signed values are used because for either map the upper bin may have a
     // higher or lower value than the lower bin
     //
-    int32_t valueDiff = value - inputMap[ lowerBin ];
-    int32_t inputBinDiff = inputMap[ upperBin ] - inputMap[ lowerBin ];
-    int32_t outputBinDiff = outputMap[ upperBin ] - outputMap[ lowerBin ];
+    int32_t valueDiff = (int32_t)value - inputMap[ lowerBin ];
+    int32_t inputBinDiff = (int32_t)inputMap[ upperBin ] - inputMap[ lowerBin ];
+    int32_t outputBinDiff =
+        (int32_t)outputMap[ upperBin ] - outputMap[ lowerBin ];
 
     int32_t output = ( valueDiff * outputBinDiff ) / inputBinDiff;
